Fix log submission to ElasticSearch
By making a new Array of bulkSize, appending new logs, and then doing a join with \n,
we would get a lot of empty newlines prepending the actual data. ElasticSearch would
reject this with a 400 error. I changed this to make an Array of size 0 - there might
be a better solution.